add logic

+++ spring-boot-bootstrap/src/main/java/org/baeldung/web/BookController.java
@@ -1,8 +1,13 @@
 package org.baeldung.web;
 
+import java.util.List;
+
 import org.baeldung.persistence.model.Book;
 import org.baeldung.persistence.repo.BookRepository;
+import org.baeldung.web.exception.BookIdMismatchException;
+import org.baeldung.web.exception.BookNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,6 +15,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -24,25 +30,42 @@ public class BookController {
         return bookRepository.findAll();
     }
 
+    @GetMapping("/title/{bookTitle}")
+    public List<Book> findByTitle(@PathVariable String bookTitle) {
+        return bookRepository.findByTitle(bookTitle);
+    }
+
     @GetMapping("/{id}")
     public Book findOne(@PathVariable Long id) {
-        return bookRepository.findOne(id);
+        Book book = bookRepository.findOne(id);
+        if (book == null)
+            throw new BookNotFoundException();
+        return book;
     }
 
     @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
     public Book create(@RequestBody Book book) {
-        // validate
         return bookRepository.save(book);
     }
 
     @DeleteMapping("/{id}")
     public void delete(@PathVariable Long id) {
+        Book book = bookRepository.findOne(id);
+        if (book == null)
+            throw new BookNotFoundException();
         bookRepository.delete(id);
     }
 
     @PutMapping("/{id}")
     public Book updateBook(@RequestBody Book book, @PathVariable Long id) {
-        // validate
+        if (book.getId() != id) {
+            throw new BookIdMismatchException();
+        }
+        Book old = bookRepository.findOne(id);
+        if (old == null) {
+            throw new BookNotFoundException();
+        }
         return bookRepository.save(book);
     }
